Files merged. Needs refactoring
